requirements.txt
asgiref==3.6.0
captcha==0.4
Django==4.1.7
numpy==1.24.2
Pillow==9.4.0
sqlparse==0.4.3
